# optimize read_in_chunks