@Kami 2023-01-28
Awesome, thanks for this contribution.

I think it would also be good to add micro benchmarks for this function / code path to `libcloud/test/benchmarks/` so we can track performance over time, avoid regressions, etc.