@Kami 2023-01-28
I've added a basic baseline micro benchmark (to help us avoid regressions in the future) in 523838cd5e5a7f0c0dcab59651ebf4dee3fe6d1b.

I was indeed able to reproduce your results (large performance speed up). Right now this benchmark only measures CPU time and not memory usage. Measuring memory usage would be a nice addition, but that would be a bit more involved change since pytest-benchmark doesn't support doing something like that natively.

![Screenshot from 2023-01-28 12-34-51](https://user-images.githubusercontent.com/125088/215264454-cd7ea01b-9edf-4a8f-9833-8d1e1f7a00f9.png)

As you mentioned, the performance gets much worse with older implementation with larger object / input data size, but to prevent micro benchmarks for running too long, I used smaller data sizes.

Thanks again and great work :tada: